Begin formalising IC messages into their own class
This commit starts a rework wherein I want to rework IC message handling for the better.

Namely, I want it so that any IC message data can be accessed through an instance of a class specifically made for it.

The goal of this (and further) commit(s) is to separate 'technical' checks from sanity checks when it comes to handling IC messages.
This new ICMessage class, for example, shall handle whether it makes sense to receive IC messages from a character with an ID of -5, but it'll still be Courtroom's tasks to decide whether the character ID of 7 is muted or not.

Even with this rework, the IC message will still be very AO-netcode-like, but I believe later on, this will be a good first step for reworking IC messages.